ISCHIA

Documentation for ISCHIA.

ISCHIA.CooccurOutput

ISCHIA.calculate_cooccurrence_matrix

ISCHIA.calculate_cooccurrence_stats

ISCHIA.effect_sizes

ISCHIA.find_differentially_cooccurring_LR_pairs

ISCHIA.find_enriched_LR_pairs

ISCHIA.summarize_cooccur

ISCHIA.CooccurOutput — Type

A mutable struct representing the output of co-occurrence analysis.

Fields

results::DataFrame: A DataFrame containing analysis results.

positive::Int: The count of positive results.

negative::Int: The count of negative results.

co_occurrences::Int: The count of co-occurrences.

pairs::Int: The total number of pairs.

random::Int: The count of random results.

unclassifiable::Int: The count of unclassifiable results.

sites::Matrix{Int}: A matrix representing site information.

species::Int: The count of species.

percent_sig::Float64: The percentage of significant results.

spp_key::Union{Nothing, DataFrame}: A DataFrame containing species keys (optional).

spp_names::Union{Vector{Int}, Vector{String}}: A vector of species names. If names weren't provided, then it would contain numerical identifiers.

omitted::Union{Nothing, Int}: The count of omitted results (optional).

pot_pairs::Union{Nothing, Int}: The count of potential pairs (optional).

true_rand_classifier::Float64: The true random classifier value.

This mutable struct is used to encapsulate and organize the results of co-occurrence analysis.

ISCHIA.calculate_cooccurrence_matrix — Method

Calculate the co-occurrence matrix N from a binary species-site matrix.

This function creates a species by species matrix of potential co-occurring sites (N) from a binary species by site matrix, where 1 represents potential occupancy, and 0 indicates species absence.

Arguments

mat::Matrix{Int}: A binary species by site matrix.

Returns

A species by species matrix where the upper triangle contains N for each species pair.

Examples

```julia

Define a binary species by site matrix

speciesmatrix = rand(Bool, numspecies, num_sites)

Calculate the co-occurrence matrix N

cooccurrencematrix = createNmatrix(speciesmatrix)

ISCHIA.calculate_cooccurrence_stats — Method

Calculate co-occurrence statistics and probabilities.

Arguments

mat::Matrix{Int}: A binary species by site matrix.

specie_names::Vector{String}: Names of species.

type::String: Type of matrix ('sppsite' or 'sitespp').

thresh::Bool: Whether to apply a threshold.

spp_names::Bool: Whether to include species names.

true_rand_classifier::Float64: True random classifier.

prob::String: Probability calculation method ('hyper' or 'comb').

site_mask::Union{Nothing, Matrix{Int}}: Matrix specifying sites.

only_effects::Bool: Whether to calculate only effects.

eff_standard::Bool: Whether to standardize effects.

eff_matrix::Bool: Whether to calculate effect matrix.

Returns

A dictionary containing various co-occurrence statistics and results.

ISCHIA.effect_sizes — Method

effect_sizes(cooccur; standardized=true)

Calculate effect sizes for co-occurrence analysis results.

Arguments

cooccur: A co-occurrence analysis object.

standardized: A boolean indicating whether to calculate standardized effect sizes.

Returns

Effect sizes as a data frame.

Example

```julia effectsizes(cooccurobject)

ISCHIA.find_differentially_cooccurring_LR_pairs — Method

Find LR (Ligand Receptor) pairs that are significantly co-occurring in one group and not in the other group.

Arguments

group1_results: Results from the EnrichedLRs function for Group 1.

group2_results: Results from the EnrichedLRs function for Group 2.

group1_max_pval: Maximum p-value threshold for significance levels of co-occurring LR pairs in Group 1.

group2_min_pval: Minimum p-value threshold for non-significance levels of co-occurring LR pairs in Group 2.

Returns

List of LR pairs enriched in Group 1 and not in Group 2.

Example

```julia result = finddifferentiallycooccurringLRpairs(resultsgroup1, resultsgroup2, 0.05, 0.1)

ISCHIA.find_enriched_LR_pairs — Method

Calculate significant co-occurring Ligand-Receptor pairs.

This function calculates co-occurring Ligand-Receptor (LR) pairs that are statistically significant based on expression levels and correlations in a spatial dataset.

Parameters:

adata::AnnData: The (spatial) anndata dataset containing expression data.

COI::Vector{String}: Cluster of Interest, a subset of spots to focus on.

Condition::Vector{String}: Condition of interest within the dataset.

LR_list::Vector{String}: List of ligands and receptors to consider.

LR_pairs::Vector{String}: List of LR pairs to analyze.

exp_th::Real: Expression threshold for binarizing the expression matrix.

corr_th::Real: Correlation threshold for LR pairs.

Returns: A dictionary containing:

"enriched_LRs": DataFrame of enriched LR pairs.

"cooccurrence_table": Co-occurrence analysis results.

ISCHIA.summarize_cooccur — Method

Summarize the results of a co-occurrence analysis.

Arguments

cooccur_output::CooccurOutput: Co-occurrence analysis object.

Returns

A summary of the co-occurrence analysis results as a dictionary.

Example

```julia summary = summarizecooccur(cooccurobject)